fix: build the modifier mask on Windows too
The CEF event flag is a C int, signed there and unsigned on Linux and
macOS, so the mask needs a cast that clippy calls redundant on the other
two. Cast, and allow the lint where it fires.

+++ shell/src/main.rs
@@ -90,9 +90,13 @@ type OsEvent<'a> = *mut u8;
 
 /// Whether the page is given first refusal on a chord: every one carrying the
 /// primary modifier: Ctrl on Linux and Windows, Cmd on macOS.
+// The cast is required on Windows, where the flag's C int is signed, and
+// redundant on Linux and macOS, where it is not: the lint has to go, because
+// neither spelling alone builds on all three.
+#[allow(clippy::unnecessary_cast)]
 fn page_first(modifiers: u32) -> bool {
-    const PRIMARY: u32 =
-        cef_event_flags_t::EVENTFLAG_CONTROL_DOWN.0 | cef_event_flags_t::EVENTFLAG_COMMAND_DOWN.0;
+    const PRIMARY: u32 = cef_event_flags_t::EVENTFLAG_CONTROL_DOWN.0 as u32
+        | cef_event_flags_t::EVENTFLAG_COMMAND_DOWN.0 as u32;
     modifiers & PRIMARY != 0
 }
